+++ FirebaseRemoteConfig/Tests/Unit/RCNRemoteConfigTest.m
@@ -171,7 +171,8 @@ typedef NS_ENUM(NSInteger, RCNTestRCInstance) {
                                                       namespace:currentNamespace
                                                       DBManager:_DBManager
                                                   configContent:configContent
-                                                      analytics:nil]);
+                                                      analytics:nil
+                                                      abtesting:nil]);
 
     _configInstances[i] = config;
     RCNConfigSettings *settings =
@@ -493,7 +494,8 @@ typedef NS_ENUM(NSInteger, RCNTestRCInstance) {
                                                       namespace:currentNamespace
                                                       DBManager:_DBManager
                                                   configContent:configContent
-                                                      analytics:nil]);
+                                                      analytics:nil
+                                                      abtesting:nil]);
 
     _configInstances[i] = config;
     RCNConfigSettings *settings =
@@ -616,7 +618,8 @@ typedef NS_ENUM(NSInteger, RCNTestRCInstance) {
                                                       namespace:currentNamespace
                                                       DBManager:_DBManager
                                                   configContent:configContent
-                                                      analytics:nil]);
+                                                      analytics:nil
+                                                      abtesting:nil]);
 
     _configInstances[i] = config;
     RCNConfigSettings *settings =
@@ -740,7 +743,8 @@ typedef NS_ENUM(NSInteger, RCNTestRCInstance) {
                                                       namespace:currentNamespace
                                                       DBManager:_DBManager
                                                   configContent:configContent
-                                                      analytics:nil]);
+                                                      analytics:nil
+                                                      abtesting:nil]);
 
     _configInstances[i] = config;
     RCNConfigSettings *settings =